Switch links over to jmespath.org
Now that we have jmespath.org, the readthedocs links will just
contain docs specific to the python library.  There'a also
a deprecation notice over at
http://jmespath.readthedocs.org/en/latest/specification.html
and I will eventually remove the spec on readthedocs entirely.

Also, the python jmespath implementation was moved over to the
jmespath org and renamed to jmespath.py so I've updated that
link as well.